small fixes

+++ build.sh
@@ -12,8 +12,10 @@ if [ -z ${SAMBA_VERSION+x} ] || [ -z ${SAMBA_VERSION+x} ]; then
   export ALPINE_VERSION=$(docker run --rm -ti "$IMG" cat /etc/alpine-release | tail -n1 | tr -d '\r')
 fi
 
-echo "check if image was already build and pushed - skip check on release version"
-echo "$@" | grep -v "release" && docker pull "$IMG:a$ALPINE_VERSION-s$SAMBA_VERSION" 2>/dev/null >/dev/null && echo "image already build" && exit 1
+if echo "$@" | grep -v "force" 2>/dev/null >/dev/null; then
+  echo "check if image was already build and pushed - skip check on release version"
+  echo "$@" | grep -v "release" && docker pull "$IMG:a$ALPINE_VERSION-s$SAMBA_VERSION" 2>/dev/null >/dev/null && echo "image already build" && exit 1
+fi
 
 docker buildx build -q --pull --no-cache --platform "$PLATFORM" -t "$IMG:a$ALPINE_VERSION-s$SAMBA_VERSION" --push .
